+++ src/internal/analytics/__tests__/mocks.ts
@@ -24,3 +24,23 @@ export function mockFunnelMetrics() {
     externalLinkInteracted: jest.fn(),
   });
 }
+
+export function mockInnerText() {
+  if (!('innerText' in HTMLElement.prototype)) {
+    // JSDom does not support the `innerText` property. For tests, `textContent` is usually close enough.
+
+    beforeEach(() =>
+      Object.defineProperty(HTMLElement.prototype, 'innerText', {
+        get() {
+          return this.textContent;
+        },
+        set(v) {
+          this.textContent = v;
+        },
+        configurable: true,
+      })
+    );
+
+    afterEach(() => delete (HTMLElement.prototype as Partial<HTMLElement>).innerText);
+  }
+}
